♻️ Extract @planner/ai package from @planner/tui
Move chat service, tools, and prompt builder into a new UI-agnostic
@planner/ai package. Dependency graph is now cli → tui → ai → core,
letting any frontend consume AI features without Ink/React.
