+++ packages/ai/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@planner/tui",
+  "name": "@planner/ai",
   "version": "0.1.0",
   "type": "module",
   "main": "./dist/index.js",
@@ -16,12 +16,7 @@
     "lint": "tsc --noEmit"
   },
   "dependencies": {
-    "@planner/ai": "workspace:*",
     "@planner/core": "workspace:*",
-    "ink": "^5.2.1",
-    "ink-select-input": "^6.2.0",
-    "ink-spinner": "^5.0.0",
-    "ink-text-input": "^6.0.0",
-    "react": "^18.3.1"
+    "openai": "^4.104.0"
   }
 }
