+++ packages/ai/src/index.ts
@@ -1,0 +1,4 @@
+export { ChatService, type StreamCallbacks } from './services/chat.service.js';
+export { getToolDefinitions, executeTool, type ToolServices } from './services/chat-tools.js';
+export { buildSystemPrompt } from './services/chat-prompt.js';
+export { createAiContainer, type AiContainer } from './container.js';
