+++ packages/ai/src/services/chat-tools.ts
