+++ packages/ai/tsconfig.json
@@ -3,13 +3,10 @@
   "compilerOptions": {
     "rootDir": "src",
     "outDir": "dist",
-    "composite": true,
-    "jsx": "react-jsx",
-    "jsxImportSource": "react"
+    "composite": true
   },
   "include": ["src"],
   "references": [
-    { "path": "../core" },
-    { "path": "../ai" }
+    { "path": "../core" }
   ]
 }
