+++ packages/tui/src/container.ts
@@ -1,26 +1,7 @@
-import {
-  getContainer,
-  type CoreContainer,
-  ConversationRepository,
-  MessageRepository,
-  getDb,
-} from '@planner/core';
-import { ChatService } from './services/chat.service.js';
+import { createAiContainer, type AiContainer } from '@planner/ai';
 
-export interface TuiContainer extends CoreContainer {
-  chatService: ChatService;
-}
+export type TuiContainer = AiContainer;
 
 export function createTuiContainer(): TuiContainer {
-  const core = getContainer();
-  const db = getDb();
-  const conversationRepo = new ConversationRepository(db);
-  const messageRepo = new MessageRepository(db);
-
-  const chatService = new ChatService(
-    conversationRepo, messageRepo, core.configService,
-    core.areaService, core.goalService, core.taskService, core.habitService, core.contextService,
-  );
-
-  return { ...core, chatService };
+  return createAiContainer();
 }
