+++ packages/tui/src/index.ts
@@ -1,5 +1,5 @@
 export { renderApp } from './tui/app.js';
 export { createTuiContainer, type TuiContainer } from './container.js';
-export { ChatService, type StreamCallbacks } from './services/chat.service.js';
-export { getToolDefinitions, executeTool, type ToolServices } from './services/chat-tools.js';
-export { buildSystemPrompt } from './services/chat-prompt.js';
+export { ChatService, type StreamCallbacks } from '@planner/ai';
+export { getToolDefinitions, executeTool, type ToolServices } from '@planner/ai';
+export { buildSystemPrompt } from '@planner/ai';
