+++ tsconfig.json
@@ -2,6 +2,7 @@
   "files": [],
   "references": [
     { "path": "packages/core" },
+    { "path": "packages/ai" },
     { "path": "packages/cli" },
     { "path": "packages/tui" }
   ]
